requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
anyio==3.7.1
async-timeout==4.0.2
attrs==23.1.0
Authlib==1.2.1
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
confection==0.1.0
cryptography==41.0.2
cymem==2.0.7
dataclasses-json==0.5.12
decorator==5.1.1
dill==0.3.6
dnspython==2.4.0
elastic-transport==8.4.0
elasticsearch==8.8.2
exceptiongroup==1.1.2
faiss-cpu==1.7.4
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.70.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
grpcio==1.56.2
grpcio-tools==1.56.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
huggingface-hub==0.16.4
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
jaraco.context==4.3.0
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
langchain==0.0.74
langchainplus-sdk==0.0.20
langcodes==3.3.0
loguru==0.7.0
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
mdurl==0.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pathy==0.10.2
Pillow==9.5.0
pinecone-client==2.2.2
preshed==3.0.8
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.11
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
PyMuPDF==1.22.5
pyparsing==3.1.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
qdrant-client==0.11.10
redis==4.6.0
referencing==0.30.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
spacy==3.6.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==1.4.49
sqlitedict==2.1.0
srsly==2.4.7
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
thinc==8.1.10
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.31.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.20.0
wasabi==1.1.2
weaviate-client==3.22.1
wikipedia==1.4.0
wolframalpha==5.0.0
xmltodict==0.13.0
yarl==1.9.2
zipp==3.16.2
